Address #24 features with keyboard-input+ branch

Closes #24

See merge request seakaart/chameleon-2!33

+++ chameleon2/main.py
@@ -398,6 +398,9 @@ class MainApp(QtWidgets.QMainWindow, QtGui.QKeyEvent, chameleon2.design.Ui_MainW
         self.setupUi(self)
         # Set up application logo on main window
         self.setWindowTitle("Chameleon 2")
+        # Enable QWidgets to capture and filter QKeyEvents
+        self.searchButton.installEventFilter(self)
+        self.listWidget.installEventFilter(self)
 
         # Differentiate sys settings between pre and post-bundling
         if getattr(sys, 'frozen', False):
@@ -901,25 +904,33 @@ class MainApp(QtWidgets.QMainWindow, QtGui.QKeyEvent, chameleon2.design.Ui_MainW
             self.progress_bar.setValue(self.progress_bar.value() + 1)
             self.progress_bar.setLabelText(f"Analyzing {mode} tag…")
 
-    def enter_key_event(self, event):
+    def eventFilter(self, obj, event):
         """
-        Allows 'Return' or 'Enter' on keyboard to be pressed in lieu of
-        clicking run button.
+        Allows installed objects to filter QKeyEvents.
 
         Parameters
         ----------
+        obj : class
+            Specific QWidget target
+
         event : class
             Event which handles keystroke input
         """
+        # Sets up filter to enable keyboard input in listWidget
+        if self.searchButton == obj and event.type() == QtCore.QEvent.KeyPress:
+            if event.key() == QtCore.Qt.Key_Tab and self.listWidget.count() > 0:
+                self.listWidget.item(0).setSelected(True)
+            elif event.key() == QtCore.Qt.Key_Tab and self.listWidget.count() == 0:
+                event.ignore()
 
-        if self.runButton.isEnabled():
-            if event.key() == QtCore.Qt.Key_Return:
-                MainApp.run_query(self)
+        # Set up filter to enable delete key within listWidget
+        if self.listWidget == obj and event.type() == QtCore.QEvent.KeyPress:
+            if event.key() == QtCore.Qt.Key_Delete and self.listWidget.count() > 0:
+                self.delete_tag()
+            elif event.key() == QtCore.Qt.Key_Delete and self.listWidget.count() == 0:
+                event.ignore()
 
-        # Added delete key to an action (WIP)
-        if event.key() == QtCore.Qt.Key_Delete:
-            LOGGER.info('Delete key pressed...')
-            self.delete_tag()
+        return super(MainApp, self).eventFilter(obj, event)
 
     # Re-enable run button when function complete
     def finished(self):
